Refactor hook imports across components to enhance organization and clarity
- Updated imports to source hooks from their respective directories under "@/components".
- Introduced new hooks for fetching project-related data, including project analysis runs and conversation counts.
- Removed the deprecated query module to streamline functionality and improve maintainability.
