+++ echo/frontend/src/components/conversation/AutoSelectConversations.tsx
@@ -1,9 +1,6 @@
-import {
-  useConversationsByProjectId,
-  useProjectById,
-  useProjectChatContext,
-} from "@/lib/query";
-import { useAddChatContextMutation, useDeleteChatContextMutation } from "./hooks";
+import { useAddChatContextMutation, useConversationsByProjectId, useDeleteChatContextMutation } from "./hooks";
+import { useProjectChatContext } from "@/components/chat/hooks";
+import { useProjectById } from "@/components/project/hooks";
 import { Trans } from "@lingui/react/macro";
 import {
   Box,
