+++ echo/frontend/src/components/layout/ProjectConversationLayout.tsx
@@ -1,9 +1,9 @@
 import { t } from "@lingui/core/macro";
-import { useConversationById } from "@/lib/query";
 import { Stack, Title } from "@mantine/core";
 import { useParams } from "react-router-dom";
 import { TabsWithRouter } from "./TabsWithRouter";
 import { ConversationStatusIndicators } from "../conversation/ConversationAccordion";
+import { useConversationById } from "../conversation/hooks";
 
 export const ProjectConversationLayout = () => {
   const { conversationId } = useParams();
