+++ echo/frontend/src/components/project/ProjectAnalysisRunStatus.tsx
@@ -1,10 +1,10 @@
 import { Trans } from "@lingui/react/macro";
 import { directus } from "@/lib/directus";
-import { useLatestProjectAnalysisRunByProjectId } from "@/lib/query";
 import { readItems } from "@directus/sdk";
 import { Alert, Stack } from "@mantine/core";
 import { useQuery } from "@tanstack/react-query";
 import { CloseableAlert } from "../common/ClosableAlert";
+import { useLatestProjectAnalysisRunByProjectId } from "./hooks";
 
 export const ProjectAnalysisRunStatus = ({
   projectId,
