+++ echo/frontend/src/hooks/useAuthenticated.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useLocation, useSearchParams } from "react-router-dom";
 import { directus } from "@/lib/directus";
-import { useLogoutMutation } from "@/lib/query";
+import { useLogoutMutation } from "@/components/auth/hooks";
 
 export const useAuthenticated = (doRedirect = false) => {
   const logoutMutation = useLogoutMutation();
